+++ ChangeLog.md
@@ -73,6 +73,7 @@ If you have any feedback [use our discussion forum](https://github.com/pasdoc/pa
     * Parse `noreturn` (Michalis)
     * Allow `on` and `out` as regular names (e.g. for properties) (Michalis)
     * Parse `align(N)` directive at record end (Michalis)
+    * Parse generics with multiple constraints separated by semicolons, like `procedure Foo<A; B>` (Michalis)
 
 * Other fixes:
     * Fixed using introduction `@shortTitle` (Michalis)